requirements.txt
openai-whisper>=20231117
torch>=2.0.0
numpy>=1.20.0
soundfile>=0.12.1
tqdm>=4.65.0
